+++ SendBirdDesk.framework/Headers/SBDSKTypes.h
@@ -1,0 +1,19 @@
+//
+//  SBDSKTypes.h
+//  SendBirdDesk
+//
+//  Created by SendBird on 11/21/17.
+//  Copyright © 2019 SendBird. All rights reserved.
+//
+
+#ifndef SBDSKTypes_h
+#define SBDSKTypes_h
+
+typedef NS_ENUM(NSInteger, SBDSKTicketPriority) {
+    SBDSKTicketPriorityLow = 0,
+    SBDSKTicketPriorityMedium = 1,
+    SBDSKTicketPriorityHigh = 2,
+    SBDSKTicketPriorityUrgent = 3
+};
+
+#endif /* SBDSKTypes_h */